refactor: move the server:search into a new datasource:search module
Move to a new module so it fits other data sources. Later on a repository with some cached data can
be used to optimize the search a little bit
