+++ app/build.gradle
@@ -63,6 +63,7 @@ dependencies {
     implementation project(":features:timeline")
     implementation project(":libraries:core")
     implementation project(":libraries:database")
+    implementation project(":libraries:datasource:search")
     implementation project(":libraries:datasource:series")
     implementation project(":libraries:datasource:trending")
     implementation project(":libraries:datasource:user")
@@ -74,7 +75,6 @@ dependencies {
     implementation project(":libraries:kitsu:trending")
     implementation project(":libraries:kitsu:user")
     implementation project(":libraries:server:auth")
-    implementation project(":libraries:server:search")
 
     implementation "androidx.appcompat:appcompat:$appcompat_version"
     implementation "androidx.constraintlayout:constraintlayout:$constraint_layout_version"
