+++ app/src/androidTest/java/com/chesire/nekome/features/search/ResultsTests.kt
@@ -2,12 +2,12 @@ package com.chesire.nekome.features.search
 
 import com.chesire.nekome.UITest
 import com.chesire.nekome.core.Resource
+import com.chesire.nekome.datasource.search.remote.SearchApi
 import com.chesire.nekome.helpers.creation.createSearchDomain
 import com.chesire.nekome.injection.SearchModule
 import com.chesire.nekome.robots.activity
 import com.chesire.nekome.robots.search.results
 import com.chesire.nekome.robots.search.search
-import com.chesire.nekome.search.api.SearchApi
 import dagger.hilt.android.testing.BindValue
 import dagger.hilt.android.testing.HiltAndroidTest
 import dagger.hilt.android.testing.UninstallModules
