+++ features/search/build.gradle
@@ -33,8 +33,8 @@ android {
 
 dependencies {
     implementation project(":libraries:core")
+    implementation project(":libraries:datasource:search")
     implementation project(":libraries:datasource:series")
-    implementation project(":libraries:server:search")
 
     implementation "androidx.appcompat:appcompat:$appcompat_version"
     implementation "androidx.constraintlayout:constraintlayout:$constraint_layout_version"
