+++ features/search/src/main/java/com/chesire/nekome/app/search/SearchViewModel.kt
@@ -6,8 +6,8 @@ import androidx.lifecycle.viewModelScope
 import com.chesire.nekome.app.search.domain.SearchDomainMapper
 import com.chesire.nekome.core.Resource
 import com.chesire.nekome.core.flags.SeriesType
-import com.chesire.nekome.search.api.SearchApi
-import com.chesire.nekome.search.api.SearchDomain
+import com.chesire.nekome.datasource.search.SearchDomain
+import com.chesire.nekome.datasource.search.remote.SearchApi
 import com.hadilq.liveevent.LiveEvent
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
