+++ libraries/datasource/search/.gitignore
