+++ libraries/datasource/search/src/main/java/com/chesire/nekome/datasource/search/remote/SearchApi.kt
@@ -1,6 +1,7 @@
-package com.chesire.nekome.search.api
+package com.chesire.nekome.datasource.search.remote
 
 import com.chesire.nekome.core.Resource
+import com.chesire.nekome.datasource.search.SearchDomain
 
 /**
  * Methods relating to searching for series.
