--- libraries/server/search/src/main/AndroidManifest.xml
@@ -1,1 +1,0 @@
-<manifest package="com.chesire.nekome.search.api" />
