+++ settings.gradle
@@ -9,6 +9,7 @@ include ':app',
 
     ':libraries:core',
     ':libraries:database',
+    ':libraries:datasource:search',
     ':libraries:datasource:series',
     ':libraries:datasource:trending',
     ':libraries:datasource:user',
@@ -20,6 +21,5 @@ include ':app',
     ':libraries:kitsu:trending',
     ':libraries:kitsu:user',
     ':libraries:server:auth',
-    ':libraries:server:search',
 
     ':testing'
